Added incident serializer and deserializer.
Created new IncidentSerializer and IncidentDeserializer classes, which allow us to
customize the conversion between JSON and Incident objects. The serializer is used
to create a json string representing the incident object, which can be sent
through Kafka to the Incident Dashboard. The deserializer is used to create a
Incident object in the create incident controller, letting an agent use a custom json format to post the incident info.

Signed-off-by: alejgh <alejandrgh11@gmail.com>